+++ docs/app/layout.config.tsx
@@ -1,8 +1,8 @@
-import type { BaseLayoutProps } from 'fumadocs-ui/layouts/shared';
+import type { BaseLayoutProps } from "fumadocs-ui/layouts/shared";
 
 export const baseOptions: BaseLayoutProps = {
   nav: {
     title: <code className="text-sm">🔗 next-validate-link</code>,
   },
-  githubUrl: 'https://github.com/fuma-nama/next-validate-link',
+  githubUrl: "https://github.com/fuma-nama/next-validate-link",
 };