+++ docs/content/docs/example.ts
@@ -1,24 +1,24 @@
+import path from "node:path";
+import { getTableOfContents } from "fumadocs-core/server";
+import { getSlugs, parseFilePath } from "fumadocs-core/source";
 import {
   printErrors,
   readFiles,
   scanURLs,
   validateFiles,
-} from 'next-validate-link';
-import { getSlugs, parseFilePath } from 'fumadocs-core/source';
-import { getTableOfContents } from 'fumadocs-core/server';
-import path from 'node:path';
+} from "next-validate-link";
 
 async function checkLinks() {
   // we read them all at once to avoid repeated file read
-  const docsFiles = await readFiles('content/docs/**/*.{md,mdx}');
+  const docsFiles = await readFiles("content/docs/**/*.{md,mdx}");
 
   // other collections too!
-  const blogFiles = await readFiles('content/blog/**/*.{md,mdx}');
+  const blogFiles = await readFiles("content/blog/**/*.{md,mdx}");
 
   const scanned = await scanURLs({
     populate: {
-      '(home)/blog/[slug]': blogFiles.map((file) => {
-        const info = parseFilePath(path.relative('content/blog', file.path));
+      "(home)/blog/[slug]": blogFiles.map((file) => {
+        const info = parseFilePath(path.relative("content/blog", file.path));
 
         return {
           value: getSlugs(info)[0],
@@ -27,8 +27,8 @@ async function checkLinks() {
           ),
         };
       }),
-      'docs/[[...slug]]': docsFiles.map((file) => {
-        const info = parseFilePath(path.relative('content/docs', file.path));
+      "docs/[[...slug]]": docsFiles.map((file) => {
+        const info = parseFilePath(path.relative("content/docs", file.path));
 
         return {
           value: getSlugs(info),