+++ src/sample.ts
@@ -1,14 +1,14 @@
-import path from 'node:path';
-import fs from 'node:fs/promises';
-import FastGlob, { type Pattern } from 'fast-glob';
-import type { FileObject } from './validate';
-import matter from 'gray-matter';
+import fs from "node:fs/promises";
+import path from "node:path";
+import matter from "gray-matter";
+import type { FileObject } from "./validate";
+import { glob } from "tinyglobby";
 
-export type PathToUrl = (path: string) => string;
+export type PathToUrl = (path: string) => string | undefined;
 
 export async function readFileFromPath(
   file: string,
-  pathToUrl?: PathToUrl,
+  pathToUrl?: PathToUrl
 ): Promise<FileObject> {
   const content = await fs
     .readFile(path.resolve(file))
@@ -21,29 +21,29 @@ export async function readFileFromPath(
     data: parsed.data,
     // apply offset to ensure line numbers are correct
     content:
-      '\n'.repeat(countLine(content) - countLine(parsed.content)) +
+      "\n".repeat(countLine(content) - countLine(parsed.content)) +
       parsed.content,
     url: pathToUrl ? pathToUrl(file) : undefined,
   };
 }
 
 export async function readFiles(
-  patterns: Pattern | Pattern[],
+  patterns: string | readonly string[],
   options: Partial<{
     pathToUrl?: PathToUrl;
-  }> = {},
+  }> = {}
 ): Promise<FileObject[]> {
-  const files = await FastGlob(patterns);
+  const files = await glob(patterns);
 
   return await Promise.all(
-    files.map((file) => readFileFromPath(file, options.pathToUrl)),
+    files.map((file) => readFileFromPath(file, options.pathToUrl))
   );
 }
 
 function countLine(s: string) {
   let out = 0;
   for (const c of s) {
-    if (c === '\n') out++;
+    if (c === "\n") out++;
   }
 
   return out;