+++ src/scan.ts
@@ -1,6 +1,6 @@
-import * as Next from './presets/next';
-import * as Astro from './presets/astro';
-import * as Nuxt from './presets/nuxt';
+import * as Astro from "./presets/astro";
+import * as Next from "./presets/next";
+import * as Nuxt from "./presets/nuxt";
 
 export type PopulateParams = Record<
   string,
@@ -46,10 +46,10 @@ export async function scanURLs({
   /**
    * @default next
    */
-  preset?: 'next' | 'astro' | 'nuxt';
+  preset?: "next" | "astro" | "nuxt";
 } = {}): Promise<ScanResult> {
-  if (preset === 'astro') return Astro.scanURLs(options);
-  if (preset === 'nuxt') return Nuxt.scanURLs(options);
+  if (preset === "astro") return Astro.scanURLs(options);
+  if (preset === "nuxt") return Nuxt.scanURLs(options);
 
   return Next.scanURLs(options);
 }