+++ test/validate.test.ts
@@ -1,29 +1,29 @@
-import { expect, test } from 'vitest';
-import { scanURLs } from '@/scan';
-import { validateFiles, type ValidateResult } from '@/validate';
-import path from 'node:path';
+import path from "node:path";
+import { expect, test } from "vitest";
+import { scanURLs } from "@/scan";
+import { type ValidateResult, validateFiles } from "@/validate";
 
 const scanned = await scanURLs({
   pages: [
-    'page.tsx',
-    'docs/page.tsx',
-    'docs/[...slug]/page.tsx',
-    'dynamic/[[...slug]]/page.tsx',
+    "page.tsx",
+    "docs/page.tsx",
+    "docs/[...slug]/page.tsx",
+    "dynamic/[[...slug]]/page.tsx",
   ],
   populate: {
-    'docs/[...slug]': [
+    "docs/[...slug]": [
       {
-        value: ['a'],
+        value: ["a"],
       },
       {
-        value: ['b'],
+        value: ["b"],
       },
       {
-        value: ['c'],
+        value: ["c"],
       },
       {
-        value: ['d'],
-        hashes: ['hash'],
+        value: ["d"],
+        hashes: ["hash"],
       },
     ],
   },
@@ -36,58 +36,58 @@ function simplify(results: ValidateResult[]) {
   }));
 }
 
-test('validate links: valid', async () => {
+test("validate links: valid", async () => {
   const pathToUrl = (file: string) => {
-    return ['docs', file.slice(0, -path.extname(file).length).split(path.sep)]
+    return ["docs", file.slice(0, -path.extname(file).length).split(path.sep)]
       .filter((v) => v.length > 0)
-      .join('/');
+      .join("/");
   };
 
   expect(
     await validateFiles(
       [
         {
-          path: 'a.md',
-          url: pathToUrl('a.md'),
-          content: '[hello](/)',
+          path: "a.md",
+          url: pathToUrl("a.md"),
+          content: "[hello](/)",
         },
         {
-          path: 'b.md',
-          url: pathToUrl('b.md'),
-          content: '[hello](/docs) [hello](/docs/d) [hello](/docs/d#hash)',
+          path: "b.md",
+          url: pathToUrl("b.md"),
+          content: "[hello](/docs) [hello](/docs/d) [hello](/docs/d#hash)",
         },
         {
-          path: 'c.md',
-          url: pathToUrl('c.md'),
+          path: "c.md",
+          url: pathToUrl("c.md"),
           content:
-            '[hello](../dynamic) [hello](../dynamic/anything) [file](./b.md)',
+            "[hello](../dynamic) [hello](../dynamic/anything) [file](./b.md)",
         },
         {
-          path: 'd.md',
-          url: pathToUrl('d.md'),
-          content: '[hello](./) [hello](./a) example@example.com',
+          path: "d.md",
+          url: pathToUrl("d.md"),
+          content: "[hello](./) [hello](./a) example@example.com",
         },
       ],
       {
         scanned,
         pathToUrl,
-      },
-    ).then(simplify),
+      }
+    ).then(simplify)
   ).toMatchInlineSnapshot(`[]`);
 });
 
-test('validate links: not found', async () => {
+test("validate links: not found", async () => {
   expect(
     await validateFiles(
       [
         {
-          path: 'a.md',
+          path: "a.md",
           content: `[hello](/docs/invalid)
 [hello](/doc)`,
         },
       ],
-      { scanned },
-    ).then(simplify),
+      { scanned }
+    ).then(simplify)
   ).toMatchInlineSnapshot(`
     [
       {
@@ -111,17 +111,17 @@ test('validate links: not found', async () => {
   `);
 });
 
-test('validate links: invalid fragments', async () => {
+test("validate links: invalid fragments", async () => {
   expect(
     await validateFiles(
       [
         {
-          path: 'a.md',
-          content: '[hello](/docs/d#invalid)',
+          path: "a.md",
+          content: "[hello](/docs/d#invalid)",
         },
       ],
-      { scanned },
-    ).then(simplify),
+      { scanned }
+    ).then(simplify)
   ).toMatchInlineSnapshot(`
     [
       {
@@ -139,21 +139,21 @@ test('validate links: invalid fragments', async () => {
   `);
 });
 
-test('validate links: external urls', async () => {
+test("validate links: external urls", async () => {
   expect(
     await validateFiles(
       [
         {
-          path: 'a.md',
-          content: 'https://google.com http://localhost:3000',
+          path: "a.md",
+          content: "https://google.com http://localhost:3000",
         },
         {
-          path: 'b.md',
-          content: 'https://invalid.com',
+          path: "b.md",
+          content: "https://invalid.com",
         },
       ],
-      { scanned, checkExternal: true },
-    ).then(simplify),
+      { scanned, checkExternal: true }
+    ).then(simplify)
   ).toMatchInlineSnapshot(`
     [
       {
@@ -171,12 +171,12 @@ test('validate links: external urls', async () => {
   `);
 });
 
-test('validate links: line numbers', async () => {
+test("validate links: line numbers", async () => {
   expect(
     await validateFiles(
       [
         {
-          path: 'a.md',
+          path: "a.md",
           content: `[line 1](/unknown)
 
 [line 3](/unknown)
@@ -186,8 +186,8 @@ test('validate links: line numbers', async () => {
 some content [line 7](/test)`,
         },
       ],
-      { scanned },
-    ).then(simplify),
+      { scanned }
+    ).then(simplify)
   ).toMatchInlineSnapshot(`
     [
       {
@@ -218,6 +218,41 @@ some content [line 7](/test)`,
           },
         ],
         "file": "a.md",
+      },
+    ]
+  `);
+});
+
+test("validate links: components", async () => {
+  expect(
+    await validateFiles(
+      [
+        {
+          path: "a.mdx",
+          content: `<Card href="/unknown">Hello World</Card>`,
+        },
+      ],
+      {
+        scanned,
+        markdown: {
+          components: {
+            Card: { attributes: ["href"] },
+          },
+        },
+      }
+    ).then(simplify)
+  ).toMatchInlineSnapshot(`
+    [
+      {
+        "errors": [
+          {
+            "column": 1,
+            "line": 1,
+            "reason": "not-found",
+            "url": "/unknown",
+          },
+        ],
+        "file": "a.mdx",
       },
     ]
   `);