+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # next-validate-link
 
+## 1.6.2
+
+### Patch Changes
+
+- 77ba13e: Switch to `tinyglobby`
+- 77ba13e: support `checkRelativePaths` & `checkRelativeUrls` options
+- 8b84c15: Support custom remark plugins
+
+## 1.6.1
+
+### Patch Changes
+
+- db5a302: Support customising Markdown validator
+
 ## 1.6.0
 
 ### Minor Changes